Added some more stuff to the decoder and a lil more to the instructions
Actually trying to do a smidge of decoding has be rethinking quite a few things, mainly the opcodes stuff. Instead of opcodes, was thinking we use masks somehow like in the decoder.

TODO: something about the opcodes and decoding

TODO: Also, what do we think about maybe moving some of the emulation functionality to inside the decoder so we dont have to then invoke the decoder afterward? Just a thought.